examples/llm_compression/torch/llama_3_2/requirements.txt
transformers
datasets==3.2
openvino==2024.6
optimum-intel[openvino]
